requirements.txt
aiogram>=3.7,<4.0
python-dotenv>=1.0,<2.0
openai>=1.0,<2.0
google-genai>=0.3,<1.0
psycopg[binary]>=3.2,<4.0
